add o arquivo unitario e alguns comentarios

+++ produtoEscalar/index.c
@@ -1,14 +1,15 @@
 #include "produtoEscalar.h"
 #define tam 2
+// 7 - Produto escalar 
 int main(){
     int x[tam],y[tam],z[tam];
-    for (size_t i = 0; i < tam; i++)
+    for (int i = 0; i < tam; i++)
     {
         printf("Digite o valor de x%d: ", i+1);
         scanf("%d", &x[i]);
         printf("Digite o valor de y%d: ", i+1);
         scanf("%d", &y[i]);
-        printf("Digite o valor de z%d: ", i+1);
+        printf("Digite o valor de z%d: ", i+1 );
         scanf("%d", &z[i]);
         
     }
